pyproject.toml
[project]
name = "airflow-rag-spike"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==3.1.0",
    "langchain-text-splitters>=0.3.11",
    "markitdown[all]>=0.1.3",
    "openai>=2.1.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.9",
    "ruff>=0.13.3",
]
